# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface Company {
  name: string;
  phone: string;
  email: string;
  address: string;
  taxInfo: string;
  logoDataUrl?: string | null;
  vatPercent: number;
  profitPercent: number;
  darkMode: boolean;
}

export interface Customer {
  id?: string;
  name: string;
  firm: string;
  phone: string;
  email: string;
  address: string;
  notes: string;
}

export interface QuoteItem {
  id?: string;
  name: string;
  quantity: number;
  unit: string;
  unitPrice: number;
}

export type QuoteData = { [key: string]: unknown };

export interface Quote {
  id?: string;
  number?: string;
  customerId?: string | null;
  customerName: string;
  title: string;
  status: string;
  total: number;
  createdAt?: string;
  data: QuoteData;
  items: QuoteItem[];
  [key: string]: unknown;
 }

export interface DataBootstrap {
  company?: Company | null;
  customers: Customer[];
  quotes: Quote[];
}

export interface HealthStatus {
  status: string;
}

export interface AnalyzeWorkRequest {
  customerName?: string;
  title: string;
  type: string;
  address?: string;
  notes?: string;
  /** @maxItems 5 */
  imageDataUrls?: string[];
}

export type AnalyzeWorkResultMaterialsItem = {
  name: string;
  quantity: number;
  unit: string;
  estimatedUnitPrice: number;
};

export interface AnalyzeWorkResult {
  workType: string;
  description: string;
  dimensions: string;
  materials: AnalyzeWorkResultMaterialsItem[];
  laborHours: number;
  approximateCost: number;
  isApproximate: boolean;
  source: string;
}