# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  name: string;
  phone: string;
  email: string;
  address: string;
  taxInfo: string;
  logoDataUrl?: string | null;
  vatPercent: number;
  profitPercent: number;
  darkMode: boolean;
}