# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Customer {
  id?: string;
  name: string;
  firm: string;
  phone: string;
  email: string;
  address: string;
  notes: string;
}