# lib/api-zod/src/generated/types/dataBootstrap.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Company } from './company';
import type { Customer } from './customer';
import type { Quote } from './quote';

export interface DataBootstrap {
  company?: Company | null;
  customers: Customer[];
  quotes: Quote[];
}