# lib/api-zod/src/generated/types/quote.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteData } from './quoteData';
import type { QuoteItem } from './quoteItem';

export interface Quote {
  id?: string;
  number?: string;
  customerId?: string | null;
  customerName: string;
  title: string;
  status: string;
  total: number;
  createdAt?: Date;
  data: QuoteData;
  items: QuoteItem[];
  [key: string]: unknown;
 }