# lib/api-zod/src/generated/types/quoteItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteItem {
  id?: string;
  name: string;
  quantity: number;
  unit: string;
  unitPrice: number;
}